requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
python-dotenv==1.0.0
requests==2.31.0
redis==5.0.1
paho-mqtt==1.6.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
